test(step5): Location 도메인 동등/대소 테스트
