+++ src/test/java/com/kkambi/racing/domain/CarTest.java
@@ -5,12 +5,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 class CarTest {
 
@@ -36,31 +34,9 @@ class CarTest {
         );
     }
 
-    @DisplayName("자동차 위치는 음수일수 없다")
-    @ValueSource(ints = {-5, -3, -1})
-    @ParameterizedTest
-    void locationMustNotBeMinus(int locationValue) {
-        // when
-        // then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Location(locationValue));
-    }
-
-    @DisplayName("자동차 위치 증가")
+    @DisplayName("자동차끼리 위치로 대소 비교할 수 있다")
     @Test
-    void locationMove() {
-        // given
-        Location location = new Location(0);
-
-        // when
-        Location locationAfterMove = location.move();
-
-        // then
-        assertThat(locationAfterMove.getValue()).isEqualTo(location.getValue() + 1);
-    }
-
-    @DisplayName("자동차끼리 위치값으로 대소 비교할 수 있다")
-    @Test
-    void compareByLocationValue() {
+    void compareByLocation() {
         // given
         Car car1 = new Car(0, "a");
         Car car2 = new Car(1, "b");
@@ -71,9 +47,9 @@ class CarTest {
         assertThat(car2).isGreaterThan(car1);
     }
 
-    @DisplayName("자동차끼리 위치값으로 동등 비교할 수 있다")
+    @DisplayName("자동차끼리 위치로 동등 비교할 수 있다")
     @Test
-    void equalsByLocationValue() {
+    void equalsByLocation() {
         // given
         Car car1 = new Car(1, "a");
         Car car2 = new Car(1, "b");
